Move Redux related files to new "store" folder - Update imports path
