+++ webapp/src/store/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import usersReducer from './users/usersSlice';
+import usersReducer from '../store/user/usersSlice';
 
 export const store = configureStore({
   reducer: {
